Fix bug in int2 encoder

+++ src/main/java/com/impossibl/postgres/system/procs/Int2s.java
@@ -62,7 +62,7 @@ public class Int2s extends SimpleProcProvider {
 			else {
 				
 				buffer.writeInt(2);
-				buffer.writeByte((Short) val);
+				buffer.writeShort((Short) val);
 			}
 			
 		}
